requirements.txt
apache-airflow[kubernetes]==1.10.9
apache-airflow==1.10.9
